refactor: rename RequestAttrs as AttributeKeys
