+++ backend/api/src/main/java/com/yat2/episode/competency/CompetencyTypeController.java
@@ -22,7 +22,7 @@ import com.yat2.episode.global.swagger.ApiErrorCodes;
 import com.yat2.episode.global.swagger.AuthRequiredErrors;
 import com.yat2.episode.mindmap.MindmapAccessValidator;
 
-import static com.yat2.episode.global.constant.RequestAttrs.USER_ID;
+import static com.yat2.episode.global.constant.AttributeKeys.USER_ID;
 
 @RequiredArgsConstructor
 @RestController
