+++ backend/api/src/main/java/com/yat2/episode/global/constant/AttributeKeys.java
@@ -1,0 +1,8 @@
+package com.yat2.episode.global.constant;
+
+public final class AttributeKeys {
+    public static final String USER_ID = "userId";
+    public static final String MINDMAP_ID = "mindmapId";
+
+    private AttributeKeys() {}
+}
