--- backend/api/src/main/java/com/yat2/episode/global/constant/RequestAttrs.java
@@ -1,7 +1,0 @@
-package com.yat2.episode.global.constant;
-
-public final class RequestAttrs {
-    public static final String USER_ID = "userId";
-
-    private RequestAttrs() {}
-}
