+++ backend/api/src/main/java/com/yat2/episode/user/UserController.java
@@ -23,7 +23,7 @@ import com.yat2.episode.global.swagger.ApiErrorCodes;
 import com.yat2.episode.global.swagger.AuthRequiredErrors;
 import com.yat2.episode.user.dto.UserMeRes;
 
-import static com.yat2.episode.global.constant.RequestAttrs.USER_ID;
+import static com.yat2.episode.global.constant.AttributeKeys.USER_ID;
 
 @RestController
 @AuthRequiredErrors
